Correciones al codigo sobre la presion

+++ Practica_MD/Densidad_radial/gr_calc.f90
@@ -5,15 +5,15 @@ integer, intent (in) :: mode
 integer(kind=8) ::  i, j, ig, k
 real(kind=8) :: dist, vb, nid
 real(kind=8), dimension(3) :: xr
-real, parameter :: Pi=3.1415927
+real(kind=8), parameter :: Pi=3.1415927
 
 
 select case (mode)
 
 case(0)
     ngr = 0
-    delg = L/(2*nhist)
-    g(:) = 0
+    delg = L/(2.0000*nhist)
+    g(:) = 0.0
 
 case(1)
     ngr = ngr + 1
@@ -33,8 +33,8 @@ case(2)
     do k = 1, nhist
         dist = delg*(k+0.5)
         vb = ((k+1)**3-k**3)*delg**3
-        nid = 4/3*Pi*vb*dens
-        g(k) = g(k)/(ngr*200*nid)
+        nid = 4.0000/3.0000*Pi*vb*dens
+        g(k) = g(k)/(ngr*200.000*nid)
     end do
 
 end select
